% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plankton_parameters.R
\name{get_plankton_parameters}
\alias{get_plankton_parameters}
\title{Get all plankton parameter data for the specified project ID and date range.}
\usage{
get_plankton_parameters(proj_id, st_dt, end_dt, token = NULL, na.rm = TRUE)
}
\arguments{
\item{proj_id}{\code{integer}: Project ID.}

\item{st_dt}{\code{character}: Start of period of interest in format 'dd.mm.yyyy'.}

\item{end_dt}{\code{character}: End of period of interest in format 'dd.mm.yyyy'.}

\item{token}{\code{character}: Valid API access token. If \code{NULL}, will first attempt to read credentials from a '.auth' file in the installation folder. If this fails, will prompt for username and password.}

\item{na.rm}{\code{logical}: Should missing values be removed? Defaults to \code{TRUE}.}
}
\value{
\code{list}: A data.frame.
}
\description{
Get all plankton parameter data for the specified project ID and date range.
}
