% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project_chemistry.R
\name{get_project_chemistry}
\alias{get_project_chemistry}
\title{Get all water chemistry data for the specified project ID and date range.}
\usage{
get_project_chemistry(proj_id, st_dt, end_dt, token = NULL, na.rm = TRUE)
}
\arguments{
\item{proj_id}{\code{integer}: Project ID.}

\item{st_dt}{\code{character}: Start of period of interest in format 'dd.mm.yyyy'.}

\item{end_dt}{\code{character}: End of period of interest in format 'dd.mm.yyyy'.}

\item{token}{\code{character}: Valid API access token. If \code{NULL}, will prompt for username and password.}

\item{na.rm}{\code{logical}: Should missing values be removed? Defaults to \code{TRUE}.}
}
\value{
\code{list}: A data.frame.
}
\description{
Get all water chemistry data for the specified project ID and date range.
}
