% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project_plankton.R
\name{get_project_plankton}
\alias{get_project_plankton}
\title{Get all plankton parameter data for the specified project ID and date range.}
\usage{
get_project_plankton(
  proj_id,
  st_dt,
  end_dt,
  param_id = NULL,
  token = NULL,
  na.rm = TRUE
)
}
\arguments{
\item{proj_id}{\code{integer}: Project ID.}

\item{st_dt}{\code{character}: Start of period of interest in format 'dd.mm.yyyy'.}

\item{end_dt}{\code{character}: End of period of interest in format 'dd.mm.yyyy'.}

\item{param_id}{\code{integer}: Optional parameter ID to filter results.}

\item{token}{\code{character}: Valid API access token. If \code{NULL}, will prompt for username and password.}

\item{na.rm}{\code{logical}: Should missing values be removed? Defaults to \code{TRUE}.}
}
\value{
\code{list}: A data.frame.
}
\description{
Get all plankton parameter data for the specified project ID and date range.
}
