% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project_stations.R
\name{get_project_stations}
\alias{get_project_stations}
\title{Get stations associated with a specific project.}
\usage{
get_project_stations(proj_id, token = NULL)
}
\arguments{
\item{proj_id}{\code{integer}: Project ID.}

\item{token}{\code{character}: Valid API access token. If \code{NULL}, will first attempt to read credentials from a '.auth' file in the installation folder. If this fails, will prompt for username and password.}
}
\value{
\code{list}: A data.frame.
}
\description{
Get stations associated with a specific project.
}
