% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_projects.R
\name{get_projects}
\alias{get_projects}
\title{Get full list of projects in the Nivadatabase/Aquamonitor.}
\usage{
get_projects(token = NULL)
}
\arguments{
\item{token}{\code{character}: Valid API access token. If \code{NULL}, will prompt for username and password.}
}
\value{
\code{list}: A data.frame.
}
\description{
Get full list of projects in the Nivadatabase/Aquamonitor.
}
