% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{login}
\alias{login}
\title{Login to Aquamonitor}
\usage{
login(username = NULL, password = NULL)
}
\arguments{
\item{username}{\code{character}: Aquamonitor username (optional, see Details).}

\item{password}{\code{character}: Aquamonitor password (optional, see Details).}
}
\value{
\code{character}: Access token valid for one day.
}
\description{
Login to Aquamonitor with your username and password.
}
\details{
For security, avoid passing the arguments directly. The function will prompt for your username and password.
}
