% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_json.R
\name{get_json}
\alias{get_json}
\title{GET JSON from the AM API'}
\usage{
get_json(token, path)
}
\arguments{
\item{token}{\code{character}: Valid API access token. If \code{NULL}, will first attempt to read credentials from a '.auth' file in the installation folder. If this fails, will prompt for username and password.}

\item{path}{\code{character}: Path to endpoint. Will be appended to 'host'.}
}
\value{
\code{list}: Server response.
}
\description{
GET JSON from the AM API'
}
