% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plankton_parameters.R
\name{get_plankton_parameters}
\alias{get_plankton_parameters}
\title{Get Plankton Parameters}
\usage{
get_plankton_parameters(name = NULL, token = NULL)
}
\arguments{
\item{name}{\code{character}: Optional parameter name to filter results. Defaults to \code{NULL}, which retrieves all parameters.}

\item{token}{\code{character}: Valid API access token. If \code{NULL}, will prompt for username and password.}
}
\value{
\code{data.frame}: A data.frame with columns: id, parameter, unit, sort.
}
\description{
Get Plankton Parameters and their descriptions from AquaMonitor
}
