% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_json.R
\name{post_json}
\alias{post_json}
\title{POST JSON to the AM API'}
\usage{
post_json(token, path, in_json)
}
\arguments{
\item{token}{\code{character}: Valid API access token. If \code{NULL}, will first attempt to read credentials from a '.auth' file in the installation folder. If this fails, will prompt for username and password.}

\item{path}{\code{character}: Path to endpoint. Will be appended to 'host'.}

\item{in_json}{\code{list}: JSON to POST.}
}
\value{
\code{list}: Server response.
}
\description{
POST JSON to the AM API'
}
